improvements on actions() and heuristic()
On actions(), removed restriction that didn't allow keeper to remove box from diamond because to pass some levels it is required.

heuristic() measures the distance between the keeper and the boxes and the boxes and the diamonds (twice in relation to the first parameter, because it is more important to have the boxes hover the diamonds!).